requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
# torch and torchvision will be installed separately for ROCm in Dockerfile
segment-anything==1.0
opencv-python-headless==4.10.0.84
Pillow==10.4.0
numpy<2.0
pydantic==2.9.2